exercises/jupyterhub-v3/requirements.txt
jupyterhub==2.3.1
jupyterhub-tmpauthenticator==0.6
jupyterhub-kubespawner==4.1.0
